ecg_python_service/requirements.txt
Flask==3.1.0
Pillow
opencv-python-headless
numpy
scipy
pypdfium2
